requirements.txt
asgiref==3.7.2
black==23.7.0
click==8.1.7
Django==4.2.4
django-debug-toolbar==4.2.0
gunicorn==21.2.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
psycopg2==2.9.7
python-decouple==3.8
sqlparse==0.4.4
